Split FrameLayout traversal modules
- Move FrameLayout argument, expression, function, return, and step traversal into focused submodules.
- Split return-body traversal by primitive and function-return families.
- Move detached frame layout tests next to their owning traversal modules.
- Remove the large frame_tests module.
