+++ src/plan/frame.rs
@@ -1,18 +1,15 @@
-use super::expression::{
-    BoolExpr, BoolFunctionExpr, CallArg, CaptureArg, Expr, FunctionExpr, FunctionFunctionExpr,
-    IntExpr, IntFunctionExpr, NilExpr, NilFunctionExpr, StringExpr, StringFunctionExpr,
-};
+mod args;
+mod expression;
+mod function;
+mod return_;
+mod step;
+
 use super::function::{Param, ParamLocal, ReturnExpr};
 use super::id::{
     BoolFunctionLocalId, BoolLocalId, FunctionFunctionLocalId, IntFunctionLocalId, IntLocalId,
     NilFunctionLocalId, NilLocalId, StringFunctionLocalId, StringLocalId,
 };
 use super::step::Step;
-use super::{
-    BoolExprKind, BoolFunctionExprKind, CallArgKind, CaptureArgKind, ExprKind, FunctionExprKind,
-    FunctionFunctionExprKind, IntExprKind, IntFunctionExprKind, NilExprKind, NilFunctionExprKind,
-    ReturnBodyKind, ReturnExprKind, StepKind, StringExprKind, StringFunctionExprKind,
-};
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 pub(crate) struct FrameLayout {
@@ -130,791 +127,93 @@ impl FrameLayout {
     pub(crate) fn function_functions(self) -> usize {
         self.function_functions
     }
+}
 
-    fn include_steps(&mut self, steps: &[Step]) {
-        for step in steps {
-            self.include_step(step);
-        }
+#[cfg(test)]
+pub(super) mod test_helpers {
+    use crate::plan::{
+        BoolFunctionExpr, BoolFunctionId, BoolFunctionValue, BoolLocalId, FunctionType,
+        IntFunctionExpr, IntFunctionId, IntFunctionValue, IntLocalId, NilFunctionExpr,
+        NilFunctionId, NilFunctionValue, NilLocalId, ParamLocal, StringFunctionExpr,
+        StringFunctionId, StringFunctionValue, StringLocalId, ValueType,
+    };
+
+    pub(super) fn int_function_expr() -> IntFunctionExpr {
+        IntFunctionExpr::value(IntFunctionValue::new(
+            IntFunctionId(0),
+            vec![ParamLocal::int(IntLocalId(0))],
+        ))
     }
 
-    fn include_step(&mut self, step: &Step) {
-        match step.kind() {
-            StepKind::LetInt { local, value, .. } => {
-                self.include_int_expr(value);
-                self.include_int(*local);
-            }
-            StepKind::LetString { local, value, .. } => {
-                self.include_string_expr(value);
-                self.include_string(*local);
-            }
-            StepKind::LetBool { local, value, .. } => {
-                self.include_bool_expr(value);
-                self.include_bool(*local);
-            }
-            StepKind::LetNil { local, value, .. } => {
-                self.include_nil_expr(value);
-                self.include_nil(*local);
-            }
-            StepKind::LetIntFunction { local, value, .. } => {
-                self.include_int_function_expr(value);
-                self.include_int_function(*local);
-            }
-            StepKind::LetStringFunction { local, value, .. } => {
-                self.include_string_function_expr(value);
-                self.include_string_function(*local);
-            }
-            StepKind::LetBoolFunction { local, value, .. } => {
-                self.include_bool_function_expr(value);
-                self.include_bool_function(*local);
-            }
-            StepKind::LetNilFunction { local, value, .. } => {
-                self.include_nil_function_expr(value);
-                self.include_nil_function(*local);
-            }
-            StepKind::LetFunctionFunction { local, value, .. } => {
-                self.include_function_function_expr(value);
-                self.include_function_function(*local);
-            }
-            StepKind::Evaluate(value) => self.include_expr(value),
-        }
+    pub(super) fn string_function_expr() -> StringFunctionExpr {
+        StringFunctionExpr::value(StringFunctionValue::new(
+            StringFunctionId(0),
+            vec![ParamLocal::string(StringLocalId(0))],
+        ))
     }
 
-    fn include_expr(&mut self, expression: &Expr) {
-        match expression.kind() {
-            ExprKind::Int(expression) => self.include_int_expr(expression),
-            ExprKind::String(expression) => self.include_string_expr(expression),
-            ExprKind::Bool(expression) => self.include_bool_expr(expression),
-            ExprKind::Nil(expression) => self.include_nil_expr(expression),
-            ExprKind::Function(expression) => self.include_function_expr(expression),
-        }
+    pub(super) fn bool_function_expr() -> BoolFunctionExpr {
+        BoolFunctionExpr::value(BoolFunctionValue::new(
+            BoolFunctionId(0),
+            vec![ParamLocal::bool(BoolLocalId(0))],
+        ))
     }
 
-    fn include_return_expr(&mut self, expression: &ReturnExpr) {
-        match expression.kind() {
-            ReturnExprKind::Int { body, .. } => self.include_int_return(body),
-            ReturnExprKind::String { body, .. } => self.include_string_return(body),
-            ReturnExprKind::Bool { body, .. } => self.include_bool_return(body),
-            ReturnExprKind::Nil { body, .. } => self.include_nil_return(body),
-            ReturnExprKind::IntFunction { body, .. } => {
-                self.include_int_function_return(body);
-            }
-            ReturnExprKind::StringFunction { body, .. } => {
-                self.include_string_function_return(body);
-            }
-            ReturnExprKind::BoolFunction { body, .. } => {
-                self.include_bool_function_return(body);
-            }
-            ReturnExprKind::NilFunction { body, .. } => {
-                self.include_nil_function_return(body);
-            }
-            ReturnExprKind::FunctionFunction { body, .. } => {
-                self.include_function_function_return(body);
-            }
-        }
+    pub(super) fn nil_function_expr() -> NilFunctionExpr {
+        NilFunctionExpr::value(NilFunctionValue::new(
+            NilFunctionId(0),
+            vec![ParamLocal::nil(NilLocalId(0))],
+        ))
     }
 
-    fn include_int_return(&mut self, body: &crate::plan::IntReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_int_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_int_return(true_);
-                self.include_int_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_int_return(branch);
-                }
-                self.include_int_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_int_return(return_);
-            }
-        }
+    pub(super) fn function_returning_int_function_type() -> FunctionType {
+        FunctionType::new(
+            Vec::new(),
+            ValueType::Function(Box::new(int_function_expr().type_().clone())),
+        )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::FrameLayout;
+    use crate::plan::{
+        BoolLocalId, IntFunctionLocalId, IntLocalId, NilFunctionLocalId, NilLocalId, ParamLocal,
+        StringLocalId,
+    };
+
+    #[test]
+    fn frame_layout_derived_surface_is_covered() {
+        let layout = FrameLayout::default();
+        let cloned = clone_value(&layout);
+
+        assert_eq!(layout, cloned);
+        assert_eq!(
+            format!("{layout:?}"),
+            "FrameLayout { ints: 0, strings: 0, bools: 0, nils: 0, int_functions: 0, string_functions: 0, bool_functions: 0, nil_functions: 0, function_functions: 0 }",
+        );
     }
 
-    fn include_string_return(&mut self, body: &crate::plan::StringReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_string_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_string_return(true_);
-                self.include_string_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_string_return(branch);
-                }
-                self.include_string_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_string_return(return_);
-            }
-        }
+    fn clone_value<T: Clone>(value: &T) -> T {
+        value.clone()
     }
 
-    fn include_bool_return(&mut self, body: &crate::plan::BoolReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_bool_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_bool_return(true_);
-                self.include_bool_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_bool_return(branch);
-                }
-                self.include_bool_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_bool_return(return_);
-            }
-        }
-    }
+    #[test]
+    fn frame_layout_includes_local_ids() {
+        let mut layout = FrameLayout::default();
 
-    fn include_nil_return(&mut self, body: &crate::plan::NilReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_nil_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_nil_return(true_);
-                self.include_nil_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_nil_return(branch);
-                }
-                self.include_nil_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_nil_return(return_);
-            }
-        }
-    }
+        layout.include_local(&ParamLocal::int(IntLocalId(1)));
+        layout.include_local(&ParamLocal::string(StringLocalId(2)));
+        layout.include_local(&ParamLocal::bool(BoolLocalId(3)));
+        layout.include_local(&ParamLocal::nil(NilLocalId(4)));
+        layout.include_int_function(IntFunctionLocalId(5));
+        layout.include_nil_function(NilFunctionLocalId(6));
 
-    fn include_int_function_return(&mut self, body: &crate::plan::IntFunctionReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_int_function_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_int_function_return(true_);
-                self.include_int_function_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_int_function_return(branch);
-                }
-                self.include_int_function_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_int_function_return(return_);
-            }
-        }
-    }
-
-    fn include_string_function_return(&mut self, body: &crate::plan::StringFunctionReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_string_function_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_string_function_return(true_);
-                self.include_string_function_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_string_function_return(branch);
-                }
-                self.include_string_function_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_string_function_return(return_);
-            }
-        }
-    }
-
-    fn include_bool_function_return(&mut self, body: &crate::plan::BoolFunctionReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_bool_function_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_bool_function_return(true_);
-                self.include_bool_function_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_bool_function_return(branch);
-                }
-                self.include_bool_function_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_bool_function_return(return_);
-            }
-        }
-    }
-
-    fn include_nil_function_return(&mut self, body: &crate::plan::NilFunctionReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_nil_function_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_nil_function_return(true_);
-                self.include_nil_function_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_nil_function_return(branch);
-                }
-                self.include_nil_function_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_nil_function_return(return_);
-            }
-        }
-    }
-
-    fn include_function_function_return(&mut self, body: &crate::plan::FunctionFunctionReturn) {
-        match body.kind() {
-            ReturnBodyKind::Expr(expression) => self.include_function_function_expr(expression),
-            ReturnBodyKind::TailCall { args, .. } => self.include_call_args(args),
-            ReturnBodyKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_function_function_return(true_);
-                self.include_function_function_return(false_);
-            }
-            ReturnBodyKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_function_function_return(branch);
-                }
-                self.include_function_function_return(fallback);
-            }
-            ReturnBodyKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_function_function_return(return_);
-            }
-        }
-    }
-
-    fn include_call_args(&mut self, args: &[CallArg]) {
-        for arg in args {
-            match arg.kind() {
-                CallArgKind::Int { value, .. } => self.include_int_expr(value),
-                CallArgKind::String { value, .. } => self.include_string_expr(value),
-                CallArgKind::Bool { value, .. } => self.include_bool_expr(value),
-                CallArgKind::Nil { value, .. } => self.include_nil_expr(value),
-                CallArgKind::IntFunction { value, .. } => self.include_int_function_expr(value),
-                CallArgKind::StringFunction { value, .. } => {
-                    self.include_string_function_expr(value);
-                }
-                CallArgKind::BoolFunction { value, .. } => self.include_bool_function_expr(value),
-                CallArgKind::NilFunction { value, .. } => self.include_nil_function_expr(value),
-                CallArgKind::FunctionFunction { value, .. } => {
-                    self.include_function_function_expr(value);
-                }
-            }
-        }
-    }
-
-    fn include_capture_args(&mut self, args: &[CaptureArg]) {
-        for arg in args {
-            match arg.kind() {
-                CaptureArgKind::Int { value, .. } => self.include_int_expr(value),
-                CaptureArgKind::String { value, .. } => self.include_string_expr(value),
-                CaptureArgKind::Bool { value, .. } => self.include_bool_expr(value),
-                CaptureArgKind::Nil { value, .. } => self.include_nil_expr(value),
-                CaptureArgKind::IntFunction { value, .. } => self.include_int_function_expr(value),
-                CaptureArgKind::StringFunction { value, .. } => {
-                    self.include_string_function_expr(value);
-                }
-                CaptureArgKind::BoolFunction { value, .. } => {
-                    self.include_bool_function_expr(value)
-                }
-                CaptureArgKind::NilFunction { value, .. } => self.include_nil_function_expr(value),
-                CaptureArgKind::FunctionFunction { value, .. } => {
-                    self.include_function_function_expr(value);
-                }
-            }
-        }
-    }
-
-    fn include_int_expr(&mut self, expression: &IntExpr) {
-        match expression.kind() {
-            IntExprKind::Value(_) => {}
-            IntExprKind::LocalGet { local, .. } => self.include_int(*local),
-            IntExprKind::Call { args, .. } => self.include_call_args(args),
-            IntExprKind::FunctionCall { function, args } => {
-                self.include_int_function_expr(function);
-                self.include_call_args(args);
-            }
-            IntExprKind::Add { left, right }
-            | IntExprKind::Sub { left, right }
-            | IntExprKind::Mult { left, right }
-            | IntExprKind::Div { left, right }
-            | IntExprKind::Remainder { left, right } => {
-                self.include_int_expr(left);
-                self.include_int_expr(right);
-            }
-            IntExprKind::Negate(value) => self.include_int_expr(value),
-            IntExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_int_expr(true_);
-                self.include_int_expr(false_);
-            }
-            IntExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_int_expr(branch);
-                }
-                self.include_int_expr(fallback);
-            }
-            IntExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_int_expr(return_);
-            }
-        }
-    }
-
-    fn include_string_expr(&mut self, expression: &StringExpr) {
-        match expression.kind() {
-            StringExprKind::Value(_) => {}
-            StringExprKind::LocalGet { local, .. } => self.include_string(*local),
-            StringExprKind::Call { args, .. } => self.include_call_args(args),
-            StringExprKind::FunctionCall { function, args } => {
-                self.include_string_function_expr(function);
-                self.include_call_args(args);
-            }
-            StringExprKind::Concatenate { left, right } => {
-                self.include_string_expr(left);
-                self.include_string_expr(right);
-            }
-            StringExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_string_expr(true_);
-                self.include_string_expr(false_);
-            }
-            StringExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_string_expr(branch);
-                }
-                self.include_string_expr(fallback);
-            }
-            StringExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_string_expr(return_);
-            }
-        }
-    }
-
-    fn include_bool_expr(&mut self, expression: &BoolExpr) {
-        match expression.kind() {
-            BoolExprKind::Value(_) => {}
-            BoolExprKind::LocalGet { local, .. } => self.include_bool(*local),
-            BoolExprKind::Call { args, .. } => self.include_call_args(args),
-            BoolExprKind::FunctionCall { function, args } => {
-                self.include_bool_function_expr(function);
-                self.include_call_args(args);
-            }
-            BoolExprKind::Not(value) => self.include_bool_expr(value),
-            BoolExprKind::LtInt { left, right } => self.include_int_binary_expr(left, right),
-            BoolExprKind::LtEqInt { left, right } => self.include_int_binary_expr(left, right),
-            BoolExprKind::GtInt { left, right } => self.include_int_binary_expr(left, right),
-            BoolExprKind::GtEqInt { left, right } => self.include_int_binary_expr(left, right),
-            BoolExprKind::Equal { left, right } => self.include_binary_expr(left, right),
-            BoolExprKind::NotEqual { left, right } => self.include_binary_expr(left, right),
-            BoolExprKind::And { left, right } => self.include_bool_binary_expr(left, right),
-            BoolExprKind::Or { left, right } => self.include_bool_binary_expr(left, right),
-            BoolExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_bool_expr(true_);
-                self.include_bool_expr(false_);
-            }
-            BoolExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_bool_expr(branch);
-                }
-                self.include_bool_expr(fallback);
-            }
-            BoolExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_bool_expr(return_);
-            }
-        }
-    }
-
-    fn include_int_binary_expr(&mut self, left: &IntExpr, right: &IntExpr) {
-        self.include_int_expr(left);
-        self.include_int_expr(right);
-    }
-
-    fn include_binary_expr(&mut self, left: &Expr, right: &Expr) {
-        self.include_expr(left);
-        self.include_expr(right);
-    }
-
-    fn include_bool_binary_expr(&mut self, left: &BoolExpr, right: &BoolExpr) {
-        self.include_bool_expr(left);
-        self.include_bool_expr(right);
-    }
-
-    fn include_nil_expr(&mut self, expression: &NilExpr) {
-        match expression.kind() {
-            NilExprKind::Value => {}
-            NilExprKind::LocalGet { local, .. } => self.include_nil(*local),
-            NilExprKind::Call { args, .. } => self.include_call_args(args),
-            NilExprKind::FunctionCall { function, args } => {
-                self.include_nil_function_expr(function);
-                self.include_call_args(args);
-            }
-            NilExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_nil_expr(true_);
-                self.include_nil_expr(false_);
-            }
-            NilExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_nil_expr(branch);
-                }
-                self.include_nil_expr(fallback);
-            }
-            NilExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_nil_expr(return_);
-            }
-        }
-    }
-
-    fn include_function_expr(&mut self, expression: &FunctionExpr) {
-        match expression.kind() {
-            FunctionExprKind::Int(expression) => self.include_int_function_expr(expression),
-            FunctionExprKind::String(expression) => self.include_string_function_expr(expression),
-            FunctionExprKind::Bool(expression) => self.include_bool_function_expr(expression),
-            FunctionExprKind::Nil(expression) => self.include_nil_function_expr(expression),
-            FunctionExprKind::Function(expression) => {
-                self.include_function_function_expr(expression);
-            }
-        }
-    }
-
-    fn include_int_function_expr(&mut self, expression: &IntFunctionExpr) {
-        match expression.kind() {
-            IntFunctionExprKind::Value(_) => {}
-            IntFunctionExprKind::Closure { captures, .. } => self.include_capture_args(captures),
-            IntFunctionExprKind::LocalGet { local, .. } => self.include_int_function(*local),
-            IntFunctionExprKind::Call { args, .. } => self.include_call_args(args),
-            IntFunctionExprKind::FunctionCall { function, args, .. } => {
-                self.include_function_function_expr(function);
-                self.include_call_args(args);
-            }
-            IntFunctionExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_int_function_expr(true_);
-                self.include_int_function_expr(false_);
-            }
-            IntFunctionExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_int_function_expr(branch);
-                }
-                self.include_int_function_expr(fallback);
-            }
-            IntFunctionExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_int_function_expr(return_);
-            }
-        }
-    }
-
-    fn include_string_function_expr(&mut self, expression: &StringFunctionExpr) {
-        match expression.kind() {
-            StringFunctionExprKind::Value(_) => {}
-            StringFunctionExprKind::Closure { captures, .. } => self.include_capture_args(captures),
-            StringFunctionExprKind::LocalGet { local, .. } => {
-                self.include_string_function(*local);
-            }
-            StringFunctionExprKind::Call { args, .. } => self.include_call_args(args),
-            StringFunctionExprKind::FunctionCall { function, args, .. } => {
-                self.include_function_function_expr(function);
-                self.include_call_args(args);
-            }
-            StringFunctionExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_string_function_expr(true_);
-                self.include_string_function_expr(false_);
-            }
-            StringFunctionExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_string_function_expr(branch);
-                }
-                self.include_string_function_expr(fallback);
-            }
-            StringFunctionExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_string_function_expr(return_);
-            }
-        }
-    }
-
-    fn include_bool_function_expr(&mut self, expression: &BoolFunctionExpr) {
-        match expression.kind() {
-            BoolFunctionExprKind::Value(_) => {}
-            BoolFunctionExprKind::Closure { captures, .. } => self.include_capture_args(captures),
-            BoolFunctionExprKind::LocalGet { local, .. } => self.include_bool_function(*local),
-            BoolFunctionExprKind::Call { args, .. } => self.include_call_args(args),
-            BoolFunctionExprKind::FunctionCall { function, args, .. } => {
-                self.include_function_function_expr(function);
-                self.include_call_args(args);
-            }
-            BoolFunctionExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_bool_function_expr(true_);
-                self.include_bool_function_expr(false_);
-            }
-            BoolFunctionExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_bool_function_expr(branch);
-                }
-                self.include_bool_function_expr(fallback);
-            }
-            BoolFunctionExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_bool_function_expr(return_);
-            }
-        }
-    }
-
-    fn include_nil_function_expr(&mut self, expression: &NilFunctionExpr) {
-        match expression.kind() {
-            NilFunctionExprKind::Value(_) => {}
-            NilFunctionExprKind::Closure { captures, .. } => self.include_capture_args(captures),
-            NilFunctionExprKind::LocalGet { local, .. } => self.include_nil_function(*local),
-            NilFunctionExprKind::Call { args, .. } => self.include_call_args(args),
-            NilFunctionExprKind::FunctionCall { function, args, .. } => {
-                self.include_function_function_expr(function);
-                self.include_call_args(args);
-            }
-            NilFunctionExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_nil_function_expr(true_);
-                self.include_nil_function_expr(false_);
-            }
-            NilFunctionExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_nil_function_expr(branch);
-                }
-                self.include_nil_function_expr(fallback);
-            }
-            NilFunctionExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_nil_function_expr(return_);
-            }
-        }
-    }
-
-    fn include_function_function_expr(&mut self, expression: &FunctionFunctionExpr) {
-        match expression.kind() {
-            FunctionFunctionExprKind::Value(_) => {}
-            FunctionFunctionExprKind::Closure { captures, .. } => {
-                self.include_capture_args(captures);
-            }
-            FunctionFunctionExprKind::LocalGet { local, .. } => {
-                self.include_function_function(*local);
-            }
-            FunctionFunctionExprKind::Call { args, .. } => self.include_call_args(args),
-            FunctionFunctionExprKind::FunctionCall { function, args, .. } => {
-                self.include_function_function_expr(function);
-                self.include_call_args(args);
-            }
-            FunctionFunctionExprKind::BoolCase {
-                subject,
-                true_,
-                false_,
-            } => {
-                self.include_bool_expr(subject);
-                self.include_function_function_expr(true_);
-                self.include_function_function_expr(false_);
-            }
-            FunctionFunctionExprKind::IntCase {
-                subject,
-                clauses,
-                fallback,
-            } => {
-                self.include_int_expr(subject);
-                for (_, branch) in clauses {
-                    self.include_function_function_expr(branch);
-                }
-                self.include_function_function_expr(fallback);
-            }
-            FunctionFunctionExprKind::Block { steps, return_ } => {
-                self.include_steps(steps);
-                self.include_function_function_expr(return_);
-            }
-        }
+        assert_eq!(layout.ints(), 2);
+        assert_eq!(layout.strings(), 3);
+        assert_eq!(layout.bools(), 4);
+        assert_eq!(layout.nils(), 5);
+        assert_eq!(layout.int_functions(), 6);
+        assert_eq!(layout.nil_functions(), 7);
     }
 }
